Added: createing square, pentagon file in component folder and removed the functions inside Square.js and Pentagon.js files
